calculate closest viable spot
